% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd8ip_features_distribution}
\alias{cd8ip_features_distribution}
\title{A \link{data.frame} with feature distribution simulated from kernel-density
estimates based on the training data.}
\format{
An object of class \code{data.frame} with 100000 rows and 92 columns.
}
\usage{
cd8ip_features_distribution
}
\description{
This is based on the \code{\link[DESeq2:vst]{DESeq2::vst()}} transformation. It can be used to
compare your data with the training data, without revealing the training
data directly.
}
\keyword{datasets}
