% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd8ip_simulated_counts}
\alias{cd8ip_simulated_counts}
\title{Simulated RNA-Seq counts, an \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} object.}
\format{
An object of class \code{ExpressionSet} with 58302 rows and 80 columns.
}
\usage{
cd8ip_simulated_counts
}
\description{
Samples with \code{CD8IMMPH} column available were simulated based on the feature
values from real samples with respective CD8 immune phenotypes. Samples
without this information where simulated based on the complete feature
distribution (irrespective of the CD8 immune phenotype).
}
\keyword{datasets}
