% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst.R
\name{cd8ip_vst_dispersion_f}
\alias{cd8ip_vst_dispersion_f}
\title{Dispersion function generated on the training dataset.}
\usage{
cd8ip_vst_dispersion_f(q)
}
\description{
Dispersion function generated on the training dataset.
}
\keyword{internal}
