% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst.R
\name{compute_vst}
\alias{compute_vst}
\title{Apply \code{\link[DESeq2:vst]{DESeq2::vst()}} transformation to the data.}
\usage{
compute_vst(matr)
}
\arguments{
\item{matr}{Input raw counts matrix.}
}
\value{
A matrix with transformed values. R
Matrix rows are genes, matrix columns are samples.
}
\description{
This will use the \code{\link[DESeq2:dispersionFunction]{DESeq2::dispersionFunction()}} learned from the
dataset used for model creation.
}
\details{
If your pipeline/protocol is different than the one used for training,
you would probably need to use some sort of a batch-correction.
}
