% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{ecdf_pvals}
\alias{ecdf_pvals}
\title{Given a matrix of observations, computes p-values for out of range features.}
\usage{
ecdf_pvals(obs, reference = NULL)
}
\arguments{
\item{obs}{a matrix of observation, where every column is a variable
and every row is an observation}

\item{reference}{either a \link{data.frame} with reference variable distributions
or a \link{list} of \link[stats:ecdf]{stats::ecdf} functions computed based on reference
distribution. If \code{NULL}, \link{cd8ip_features_distribution} is used.}
}
\value{
a \link{data.frame} of p-values
}
\description{
For every observation the function will compute it's two-sided p-value
based on \code{\link[stats:ecdf]{stats::ecdf()}}. This will tell you how unlikely is it to observe
such a value given the reference distribution.Then it will use Holm's method to correct
p-values for every observation and return number of significantly
out-of-range features.
}
