% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{get_n_out_of_range}
\alias{get_n_out_of_range}
\title{Given a matrix of observations, computes how many features are out of range.}
\usage{
get_n_out_of_range(obs, reference, thr = 0.01)
}
\arguments{
\item{obs}{a matrix of observation, where every column is a variable
and every row is an observation}

\item{reference}{either a \link{data.frame} with reference variable distributions
or a \link{list} of \link[stats:ecdf]{stats::ecdf} functions computed based on reference
distribution. If \code{NULL}, \link{cd8ip_features_distribution} is used.}

\item{thr}{p-value threshold}
}
\value{
a vector with number of out-of-range features for every sample
}
\description{
See \code{\link[=ecdf_pvals]{ecdf_pvals()}} for details. It will use Holm's method to correct
p-values for every observation and return number of significantly
out-of-range features.
}
