% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cd8ip.R
\name{predict_cd8ip}
\alias{predict_cd8ip}
\title{Predict a CD8 immune phenotype using RNA-Seq data.}
\usage{
predict_cd8ip(df)
}
\arguments{
\item{df}{a \link{data.frame} with each row corresponding to a single observation.
This \link{data.frame} should have all the columns available in the
\link{cd8ip_features_distribution}. Feature values should be based on the
\code{DESeq2::vst()} transformation.}
}
\value{
a \code{\link[=data.frame]{data.frame()}} with class probabilities (\code{prob.inflamed},
\code{prob.excluded}, \code{prob.desert}) and the most-likely class
\code{predicted.class}. The \code{n_out_of_range} column indicates how many out of
range features were in the sample. In case more than 6 features
are out of range, the predictions will be \code{NA}s.
}
\description{
Predict a CD8 immune phenotype using RNA-Seq data.
}
