% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{two_sided_p}
\alias{two_sided_p}
\title{Computes a two-sided p-value from a one-sided p-value.}
\usage{
two_sided_p(p)
}
\arguments{
\item{p}{a vector of p-values}
}
\value{
a vector of two-sided p-values
}
\description{
Computes a two-sided p-value from a one-sided p-value.
}
